+++ src/components/mypage/PostPreview.tsx
